requirements.txt
customtkinter==5.2.0
pillow==10.0.1
pandas==2.1.1
pyyaml==6.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.0
keyboard==0.13.5
scikit-image==0.22.0